use Requires.jl to remove direct dependency on PyCall.jl
